Got process tunneling over SSH working.
Unfortunately, there's a race condition that I can't track in Processes. In some cases
(particularly with long IO latencies), there seems to be a rogue runnable left stuck in
the executor. A call to executor.stop() seems to clear things out, so I've added a method
on Procecsses that kludges this for you. I think we'll need a concurrency guru to fix
this one prpertly.